Added bhom base child window, and refactored dependant windows
refactored some of bhom base window logic to do this, into bhom window shell
added a modal window, that is a simple use case of this for child windows, ensure rendering of child windows
