+++ Python_Engine/Python/src/python_toolkit/bhom_tkinter/__init__.py
@@ -13,7 +13,9 @@ from .widgets import (
 	RadioSelection,
 	ValidatedEntryBox,
 )
+from .bhom_base_child_window import BHoMBaseChildWindow
 from .windows import (
+	BHoMModalWindow,
 	DirectoryFileSelector,
 	LandingPage,
 	ProcessingWindow,
@@ -41,6 +43,8 @@ __all__ = [
 	"PathSelector",
 	"RadioSelection",
 	"ValidatedEntryBox",
+	"BHoMBaseChildWindow",
+	"BHoMModalWindow",
 	"DirectoryFileSelector",
 	"LandingPage",
 	"ProcessingWindow",
