+++ Python_Engine/Python/src/python_toolkit/bhom_tkinter/windows/__init__.py
@@ -1,9 +1,13 @@
+from python_toolkit.bhom_tkinter.bhom_base_child_window import BHoMBaseChildWindow
 from .directory_file_selector import DirectoryFileSelector
 from .landing_page import LandingPage
+from .modal_window import BHoMModalWindow
 from .processing_window import ProcessingWindow
 from .warning_box import WarningBox
 
 __all__ = [
+    "BHoMBaseChildWindow",
+    "BHoMModalWindow",
     "DirectoryFileSelector",
     "LandingPage",
     "ProcessingWindow",
