+++ Python_Engine/Python/src/python_toolkit/bhom_tkinter/windows/processing_window.py
@@ -5,9 +5,9 @@ import os
 import time
 import threading
 
-from python_toolkit.bhom_tkinter.bhom_base_window import BHoMBaseWindow
+from python_toolkit.bhom_tkinter.bhom_base_child_window import BHoMBaseChildWindow
 
-class ProcessingWindow(BHoMBaseWindow):
+class ProcessingWindow(BHoMBaseChildWindow):
     """A simple processing window with animated indicator."""
 
     def __init__(self, title="Processing", message="Processing...", *args, **kwargs):
@@ -25,19 +25,18 @@ class ProcessingWindow(BHoMBaseWindow):
             theme_mode="auto",
             show_close=False,
             show_submit=False,
+            show_banner=False,
+            top_most=True,
             *args,
             **kwargs
         )
-        
-        self.title(title)
+
         self.attributes("-topmost", True)
         self.resizable(False, False)
 
-        # Container
-        container = ttk.Frame(self, padding=20)
+        container = ttk.Frame(self.content_frame, padding=20)
         container.pack(fill="both", expand=True)
 
-        # Message label (to calculate size)
         self.message_label = ttk.Label(
             container,
             text=message,
@@ -53,7 +52,6 @@ class ProcessingWindow(BHoMBaseWindow):
             pass
         self.message_label.pack(pady=(0, 20))
 
-        # Animation frame
         animation_frame = ttk.Frame(container)
         animation_frame.pack(expand=True)
 
@@ -71,30 +69,12 @@ class ProcessingWindow(BHoMBaseWindow):
             pass
         self.animation_label.pack()
 
-        # Animation state
         self.animation_frames = ["●", "●", "●"]
         self.current_frame = 0
         self.is_running = False
 
-        # Update to calculate the required size
         self.update_idletasks()
-        
-        # Get the required width and height
-        required_width = self.winfo_reqwidth()
-        required_height = self.winfo_reqheight()
-        
-        # Set minimum size
-        min_width = 300
-        min_height = 150
-        window_width = max(required_width, min_width)
-        window_height = max(required_height, min_height)
-
-        # Center on screen
-        screen_width = self.winfo_screenwidth()
-        screen_height = self.winfo_screenheight()
-        x = (screen_width - window_width) // 2
-        y = (screen_height - window_height) // 2
-        self.geometry(f"{window_width}x{window_height}+{x}+{y}")
+        self.refresh_sizing()
 
 
     def start(self):
@@ -103,7 +83,6 @@ class ProcessingWindow(BHoMBaseWindow):
             return
         self.is_running = True
 
-        # Run the Tk mainloop on the calling thread (must be main thread on many platforms).
         try:
             self._animate()
             self.mainloop()
@@ -153,24 +132,11 @@ class ProcessingWindow(BHoMBaseWindow):
     def stop(self):
         """Stop the animation and close the window."""
         self.is_running = False
-        try:
-            # Stop the mainloop if running and then destroy the window
-            if self.winfo_exists():
-                try:
-                    self.quit()
-                except Exception:
-                    pass
-                try:
-                    self.destroy()
-                except Exception:
-                    pass
-        except Exception:
-            pass
+        self.destroy_root()
 
     def _animate(self):
         """Update animation frames."""
         if self.is_running:
-            # Create rotating dot animation
             dots = ["◐", "◓", "◑", "◒"]
             self.animation_label.config(text=dots[self.current_frame % len(dots)])
             self.current_frame += 1
@@ -180,15 +146,12 @@ class ProcessingWindow(BHoMBaseWindow):
         """Update the message text."""
         try:
             self.message_label.config(text=message)
-            # schedule an idle update so the UI refreshes promptly
             self.update_idletasks()
         except Exception:
             pass
 
 
 if __name__ == "__main__":
-    # Test the processing window
-    
     processing = ProcessingWindow(title="Test Processing", message="Running Comfort and Safety Calculation...")
     def worker():
         for i in range(50):
